+++ CHANGELOG.md
@@ -45,6 +45,7 @@ This name should be decided amongst the team before the release.
 [Full list of changes](https://github.com/tweag/topiary/compare/v0.6.1...HEAD)
 
 ### Added
+- [#921](https://github.com/tweag/topiary/pull/921) man pages for the Topiary CLI
 - [#1015](https://github.com/tweag/topiary/pull/1015) Add support for include expressions to Nickel formatting
 - [#1049](https://github.com/tweag/topiary/pull/1015) Add support for rich grammar parsing errors using miette
 - [#1069](https://github.com/tweag/topiary/pull/1069) split CI nix dependent tests into separate jobs